Version 1.3.0
Removed animation for detail page and only kept connected animation. This makes navigating about 3x faster.
Made Gallery load files in batches of 50(will probably make it configurable later on) so the time to interactable is much lower than it was before. The loading indicator still needs to be removed or hidden earlier.
Other various changes...